+++ src/custom-ui-select/component/custom.ui.select.device.component.js
@@ -7,13 +7,36 @@ angular.module('opengate-angular-js').controller('customUiSelectDeviceController
         ctrl.ownConfig = {
             builder: $api().devicesSearchBuilder(),
             filter: function(search) {
-                return {
+                var filter = {
                     'or': [
                         { 'like': { 'provision.administration.identifier': search } },
                         { 'like': { 'provision.device.specificType': search } },
                         { 'like': { 'device.specificType': search } }
                     ]
                 };
+
+                if (!!ctrl.specificType) {
+                    filter = {
+                        'and': [
+                            filter,
+                            {
+                                'or': [{
+                                        'eq': {
+                                            'device.specificType': ctrl.specificType
+                                        }
+                                    },
+                                    {
+                                        'eq': {
+                                            'provision.device.specificType': ctrl.specificType
+                                        }
+                                    }
+                                ]
+                            }
+                        ]
+                    };
+                }
+
+                return filter;
             },
             rootKey: 'devices',
             collection: [],
@@ -21,14 +44,32 @@ angular.module('opengate-angular-js').controller('customUiSelectDeviceController
         };
 
         ctrl.deviceSelected = function($item, $model) {
-            var returnObj = {};
-            returnObj.$item = $item;
-            returnObj.$model = $model;
-            ctrl.onSelectItem(returnObj);
+            if (ctrl.multiple) {
+                var identifierTmp = [];
+
+                angular.forEach(ctrl.device, function(deviceTmp) {
+                    identifierTmp.push(deviceTmp.provision.administration.identifier._current.value);
+                });
+
+                ctrl.ngModel = identifierTmp;
+            } else {
+                ctrl.ngModel = $item.provision.administration.identifier._current.value;
+            }
+
+            if (ctrl.onSelectItem) {
+                var returnObj = {};
+                returnObj.$item = $item;
+                returnObj.$model = $model;
+                ctrl.onSelectItem(returnObj);
+            }
         };
 
         ctrl.deviceRemove = function($item, $model) {
-            ctrl.onRemove($item, $model);
+            if (ctrl.onRemove) {
+                ctrl.onRemove($item, $model);
+            }
+            ctrl.ngModel = null;
+
         };
 
         if (!ctrl.action) {
@@ -48,6 +89,65 @@ angular.module('opengate-angular-js').controller('customUiSelectDeviceController
                 permissions: 'manageEntity'
             };
         }
+
+        ctrl.$onChanges = function(changesObj) {
+            if (changesObj && changesObj.identifier) {
+                mapIdentifier(changesObj.identifier.currentValue);
+            }
+        };
+
+        if (ctrl.identifier) {
+            mapIdentifier(ctrl.identifier);
+        }
+
+        function mapIdentifier(identifierSource) {
+            var identifier = identifierSource;
+
+            if (identifier) {
+                if (identifier._current) {
+                    identifier = identifier._current.value;
+                }
+                if (ctrl.multiple) {
+                    if (angular.isArray(identifier)) {
+                        ctrl.device = [];
+
+                        angular.forEach(identifier, function(idTmp) {
+                            ctrl.device.push({
+                                provision: {
+                                    administration: {
+                                        identifier: {
+                                            _current: { value: idTmp }
+                                        }
+                                    },
+                                    device: {
+                                        identifier: {
+                                            _current: { value: idTmp }
+                                        }
+                                    }
+                                }
+                            })
+                        });
+                    }
+                } else {
+                    ctrl.device = [{
+                        provision: {
+                            administration: {
+                                identifier: {
+                                    _current: { value: ctrl.identifier }
+                                }
+                            },
+                            device: {
+                                identifier: {
+                                    _current: { value: ctrl.identifier }
+                                }
+                            }
+                        }
+                    }];
+                }
+            } else {
+                ctrl.device = [];
+            }
+        }
     }
 ]);
 
@@ -59,11 +159,14 @@ angular.module('opengate-angular-js').component('customUiSelectDevice', {
         onSelectItem: '&',
         onRemove: '&',
         device: '=',
+        identifier: '<?',
         multiple: '<',
         required: '=',
         label: '=',
         action: '=?',
-        specificType: '@?'
+        specificType: '@?',
+        disabled: '<?',
+        ngModel: '=?'
     }
 
 });
